+++ src/components/CurrentSnack/CurrentSnack.js
@@ -3,11 +3,11 @@ import {
   CurrentSnackContainer,
   CurrentSnackImg,
   CurrentSnackBox,
-} from "../SnackVote/SnackVoteStyled";
+} from "./CurrentSnackStyled";
 
-import "./snack.css";
+import "./CurrentSnack.css";
 
-export const SnackVote = (props) => {
+export const CurrentSnack = (props) => {
   /**
    * Thought I would try and show an example of a "dumb component" 
    * to help render the current selections
