+++ src/components/CurrentSnack/CurrentSnackStyled.js
